try supporting compose v1 and v2 for tests
